refactor: style파일 분리 => 하나의 컴포넌트로 변경
