+++ src/components/Carousel/index.tsx
@@ -1,21 +1,14 @@
 import type { ReactElement, SetStateAction, TouchEventHandler } from 'react'
-import {
-  StyledArrow,
-  StyledArrowBox,
-  StyledCarouselContainer,
-  StyledCurrentDot,
-  StyledDot,
-  StyledDotBox,
-  StyledImage,
-  StyledImageBox,
-  StyledSlider
-} from './style'
 import { useEffect, useState } from 'react'
 import { ICON } from '@constants/icons'
 
 export interface CarouselProps {
   images: { url: string; id: number }[]
   isArrow: boolean
+}
+
+interface CurrentDotProps {
+  imageIndex: number
 }
 
 const Carousel = ({ images, isArrow }: CarouselProps): ReactElement => {
@@ -125,3 +118,99 @@ const Carousel = ({ images, isArrow }: CarouselProps): ReactElement => {
 }
 
 export default Carousel
+
+import styled from '@emotion/styled'
+
+interface SliderProps {
+  cursorOn: boolean
+}
+export const StyledSlider = styled.div<SliderProps>`
+  position: relative;
+  max-width: 60vw;
+  height: 500px;
+  display: flex;
+  overflow: hidden;
+  margin: 0 auto;
+  cursor: ${({ cursorOn }): string | boolean => cursorOn && 'pointer'};
+`
+
+interface ImageBoxProps {
+  translateValue: number | null
+}
+export const StyledImageBox = styled.div<ImageBoxProps>`
+  display: flex;
+  transition: 1s;
+  transform: ${({ translateValue }): string =>
+    `translateX(-${translateValue}vw)`};
+`
+
+export const StyledImage = styled.img`
+  width: 60vw;
+  object-fit: cover;
+  object-position: center center;
+`
+
+export const StyledArrowBox = styled.div`
+  position: absolute;
+  left: -20px;
+  top: 0;
+  width: 100%;
+  height: 100%;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 20px;
+`
+
+export const StyledArrow = styled.img`
+  width: 40px;
+  height: 60px;
+  background-color: #ffffff;
+  color: white;
+  font-size: 2rem;
+  cursor: pointer;
+  &:hover {
+    color: dodgerblue;
+  }
+`
+export const StyledCarouselContainer = styled.div`
+  touch-action: none;
+  user-select: none;
+  position: relative;
+`
+
+export const StyledDotBox = styled.div`
+  display: flex;
+  gap: 5px;
+  position: absolute;
+  left: 50%;
+  bottom: -30px;
+  cursor: pointer;
+  transform: translateX(-50%);
+`
+
+export const StyledDot = styled.div`
+  width: 10px;
+  height: 10px;
+  background-color: #e8e8ea;
+  border-radius: 100px;
+  height: 10px;
+  margin: 0 1px;
+  font-size: 20px;
+`
+
+export const StyledCurrentDot = styled.div<CurrentDotProps>`
+  width: 10px;
+  height: 10px;
+  background-color: #2f2e36;
+  position: absolute;
+  left: 0;
+  top: 50%;
+  font-size: 20px;
+  margin: 0 1px;
+  border-radius: 100px;
+  cursor: pointer;
+  transform: ${({ imageIndex }): string =>
+    `translate(${imageIndex * 18}px,-50%)`};
+  transition: transform 0.5s;
+`
